fix: preserve hidden value
